Persistence code-style refactoring
