+++ src/core/Akka.Persistence/AtLeastOnceDeliveryActor.cs
@@ -1,5 +1,5 @@
 ﻿//-----------------------------------------------------------------------
-// <copyright file="AtLeastOnceDelivery.cs" company="Akka.NET Project">
+// <copyright file="AtLeastOnceDeliveryActor.cs" company="Akka.NET Project">
 //     Copyright (C) 2009-2016 Lightbend Inc. <http://www.lightbend.com>
 //     Copyright (C) 2013-2016 Akka.NET project <https://github.com/akkadotnet/akka.net>
 // </copyright>
@@ -10,7 +10,6 @@ using Akka.Actor;
 
 namespace Akka.Persistence
 {
-
     /// <summary>
     /// Persistent actor type that sends messages with at-least-once delivery semantics to destinations.
     /// It takes care of re-sending messages when they haven't been confirmed withing expected timeout.
@@ -45,16 +44,15 @@ namespace Akka.Persistence
         private readonly AtLeastOnceDeliverySemantic _atLeastOnceDeliverySemantic;
 
         /// <summary>
-        /// TBD
+        /// Initializes a new instance of the <see cref="AtLeastOnceDeliveryActor"/> class.
         /// </summary>
         protected AtLeastOnceDeliveryActor()
         {
             _atLeastOnceDeliverySemantic = new AtLeastOnceDeliverySemantic(Context, Extension.Settings.AtLeastOnceDelivery);
         }
 
-
         /// <summary>
-        /// TBD
+        /// Initializes a new instance of the <see cref="AtLeastOnceDeliveryActor"/> class.
         /// </summary>
         /// <param name="settings">TBD</param>
         protected AtLeastOnceDeliveryActor(PersistenceSettings.AtLeastOnceDeliverySettings settings)
@@ -69,10 +67,7 @@ namespace Akka.Persistence
         /// configuration key. This method can be overridden by implementation classes to return
         /// non-default values.
         /// </summary>
-        public virtual TimeSpan RedeliverInterval
-        {
-            get { return _atLeastOnceDeliverySemantic.RedeliverInterval; }
-        }
+        public virtual TimeSpan RedeliverInterval => _atLeastOnceDeliverySemantic.RedeliverInterval;
 
         /// <summary>
         /// Maximum number of unconfirmed messages that will be sent at each redelivery burst
@@ -84,10 +79,7 @@ namespace Akka.Persistence
         /// configuration key. This method can be overridden by implementation classes to return
         /// non-default values.
         /// </summary>
-        public virtual int RedeliveryBurstLimit
-        {
-            get { return _atLeastOnceDeliverySemantic.RedeliveryBurstLimit; }
-        }
+        public virtual int RedeliveryBurstLimit => _atLeastOnceDeliverySemantic.RedeliveryBurstLimit;
 
         /// <summary>
         /// After this number of delivery attempts a <see cref="UnconfirmedWarning" /> message will be sent to
@@ -97,10 +89,7 @@ namespace Akka.Persistence
         /// configuration key. This method can be overridden by implementation classes to return
         /// non-default values.
         /// </summary>
-        public int WarnAfterNumberOfUnconfirmedAttempts
-        {
-            get { return _atLeastOnceDeliverySemantic.WarnAfterNumberOfUnconfirmedAttempts; }
-        }
+        public int WarnAfterNumberOfUnconfirmedAttempts => _atLeastOnceDeliverySemantic.WarnAfterNumberOfUnconfirmedAttempts;
 
         /// <summary>
         /// Maximum number of unconfirmed messages, that this actor is allowed to hold in the memory.
@@ -111,54 +100,35 @@ namespace Akka.Persistence
         /// configuration key. This method can be overridden by implementation classes to return
         /// non-default values.
         /// </summary>
-        public int MaxUnconfirmedMessages
-        {
-            get { return _atLeastOnceDeliverySemantic.MaxUnconfirmedMessages; }
-        }
-        
+        public int MaxUnconfirmedMessages => _atLeastOnceDeliverySemantic.MaxUnconfirmedMessages;
+
         /// <summary>
         /// Number of messages that have not been confirmed yet.
         /// </summary>
-        public int UnconfirmedCount
-        {
-            get { return _atLeastOnceDeliverySemantic.UnconfirmedCount; }
-        }
+        public int UnconfirmedCount => _atLeastOnceDeliverySemantic.UnconfirmedCount;
 
-        /// <summary>
-        /// TBD
-        /// </summary>
-        /// <param name="cause">TBD</param>
-        /// <param name="message">TBD</param>
+        /// <inheritdoc />
         public override void AroundPreRestart(Exception cause, object message)
         {
             _atLeastOnceDeliverySemantic.Cancel();
             base.AroundPreRestart(cause, message);
         }
 
-        /// <summary>
-        /// TBD
-        /// </summary>
+        /// <inheritdoc />
         public override void AroundPostStop()
         {
             _atLeastOnceDeliverySemantic.Cancel();
             base.AroundPostStop();
         }
 
-        /// <summary>
-        /// TBD
-        /// </summary>
+        /// <inheritdoc />
         protected override void OnReplaySuccess()
         {
             _atLeastOnceDeliverySemantic.OnReplaySuccess();
             base.OnReplaySuccess();
         }
 
-        /// <summary>
-        /// TBD
-        /// </summary>
-        /// <param name="receive">TBD</param>
-        /// <param name="message">TBD</param>
-        /// <returns>TBD</returns>
+        /// <inheritdoc />
         protected internal override bool AroundReceive(Receive receive, object message)
         {
             return _atLeastOnceDeliverySemantic.AroundReceive(receive, message) || base.AroundReceive(receive, message);
