+++ src/core/Akka.Persistence/PersistentActor.cs
@@ -7,259 +7,18 @@
 
 using System;
 using System.Collections.Generic;
-using System.Runtime.Serialization;
 using Akka.Actor;
 using Akka.Actor.Internal;
-using Akka.Configuration;
 using Akka.Tools.MatchHandler;
 
 namespace Akka.Persistence
 {
     /// <summary>
-    /// TBD
-    /// </summary>
-    [Serializable]
-    public sealed class RecoveryCompleted
-    {
-        /// <summary>
-        /// TBD
-        /// </summary>
-        public static readonly RecoveryCompleted Instance = new RecoveryCompleted();
-        private RecoveryCompleted(){}
-
-        /// <summary>
-        /// TBD
-        /// </summary>
-        /// <param name="obj">TBD</param>
-        /// <returns>TBD</returns>
-        public override bool Equals(object obj)
-        {
-            return obj is RecoveryCompleted;
-        }
-    }
-
-    /// <summary>
-    /// Recovery mode configuration object to be return in <see cref="PersistentActor.get_Recovery()"/>
-    /// 
-    /// By default recovers from latest snashot replays through to the last available event (last sequenceNr).
-    /// 
-    /// Recovery will start from a snapshot if the persistent actor has previously saved one or more snapshots
-    /// and at least one of these snapshots matches the specified <see cref="FromSnapshot"/> criteria.
-    /// Otherwise, recovery will start from scratch by replaying all stored events.
-    /// 
-    /// If recovery starts from a snapshot, the <see cref="PersistentActor"/> is offered that snapshot with a
-    /// <see cref="SnapshotOffer"/> message, followed by replayed messages, if any, that are younger than the snapshot, up to the
-    /// specified upper sequence number bound (<see cref="ToSequenceNr"/>).
-    /// </summary>
-    [Serializable]
-    public sealed class Recovery
-    {
-        /// <summary>
-        /// TBD
-        /// </summary>
-        public static readonly Recovery Default = new Recovery(SnapshotSelectionCriteria.Latest);
-        /// <summary>
-        /// TBD
-        /// </summary>
-        public static readonly Recovery None = new Recovery(SnapshotSelectionCriteria.Latest, 0);
-
-        /// <summary>
-        /// TBD
-        /// </summary>
-        public Recovery() : this(SnapshotSelectionCriteria.Latest)
-        {
-        }
-
-        /// <summary>
-        /// TBD
-        /// </summary>
-        /// <param name="fromSnapshot">TBD</param>
-        public Recovery(SnapshotSelectionCriteria fromSnapshot) : this(fromSnapshot, long.MaxValue)
-        {
-        }
-
-        /// <summary>
-        /// TBD
-        /// </summary>
-        /// <param name="fromSnapshot">TBD</param>
-        /// <param name="toSequenceNr">TBD</param>
-        public Recovery(SnapshotSelectionCriteria fromSnapshot, long toSequenceNr) : this(fromSnapshot, toSequenceNr, long.MaxValue)
-        {
-        }
-
-        /// <summary>
-        /// TBD
-        /// </summary>
-        /// <param name="fromSnapshot">TBD</param>
-        /// <param name="toSequenceNr">TBD</param>
-        /// <param name="replayMax">TBD</param>
-        public Recovery(SnapshotSelectionCriteria fromSnapshot = null, long toSequenceNr = long.MaxValue, long replayMax = long.MaxValue)
-        {
-            FromSnapshot = fromSnapshot != null ? fromSnapshot : SnapshotSelectionCriteria.Latest;
-            ToSequenceNr = toSequenceNr;
-            ReplayMax = replayMax;
-        }
-
-        /// <summary>
-        /// Criteria for selecting a saved snapshot from which recovery should start. Default is latest (= youngest) snapshot.
-        /// </summary>
-        public SnapshotSelectionCriteria FromSnapshot { get; private set; }
-
-        /// <summary>
-        /// Upper, inclusive sequence number bound for recovery. Default is no upper bound.
-        /// </summary>
-        public long ToSequenceNr { get; private set; }
-
-        /// <summary>
-        /// Maximum number of messages to replay. Default is no limit.
-        /// </summary>
-        public long ReplayMax { get; private set; }
-    }
-
-    /// <summary>
-    /// TBD
-    /// </summary>
-    public sealed class RecoveryTimedOutException : AkkaException
-    {
-        /// <summary>
-        /// Initializes a new instance of the <see cref="RecoveryTimedOutException"/> class.
-        /// </summary>
-        public RecoveryTimedOutException()
-        {
-        }
-
-        /// <summary>
-        /// Initializes a new instance of the <see cref="RecoveryTimedOutException"/> class.
-        /// </summary>
-        /// <param name="message">The message that describes the error.</param>
-        /// <param name="cause">The exception that is the cause of the current exception.</param>
-        public RecoveryTimedOutException(string message, Exception cause = null) : base(message, cause)
-        {
-        }
-
-        /// <summary>
-        /// Initializes a new instance of the <see cref="RecoveryTimedOutException"/> class.
-        /// </summary>
-        /// <param name="info">The <see cref="SerializationInfo" /> that holds the serialized object data about the exception being thrown.</param>
-        /// <param name="context">The <see cref="StreamingContext" /> that contains contextual information about the source or destination.</param>
-        public RecoveryTimedOutException(SerializationInfo info, StreamingContext context) : base(info, context)
-        {
-        }
-    }
-
-    /// <summary>
-    /// This defines how to handle the current received message which failed to stash, when the size
-    /// of the Stash exceeding the capacity of the Stash.
-    /// </summary>
-    public interface IStashOverflowStrategy { }
-
-    /// <summary>
-    /// Discard the message to <see cref="DeadLetterActorRef"/>
-    /// </summary>
-    public class DiscardToDeadLetterStrategy : IStashOverflowStrategy
-    {
-        /// <summary>
-        /// TBD
-        /// </summary>
-        public static readonly DiscardToDeadLetterStrategy Instance = new DiscardToDeadLetterStrategy();
-
-        private DiscardToDeadLetterStrategy() { }
-    }
-
-    /// <summary>
-    /// Throw <see cref="StashOverflowException"/>, hence the persistent actor will start recovery
-    /// if guarded by default supervisor strategy.
-    /// Be careful if used together with <see cref="Eventsourced.Persist{TEvent}(TEvent,Action{TEvent})"/>
-    /// or <see cref="Eventsourced.PersistAll{TEvent}(IEnumerable{TEvent},Action{TEvent})"/>
-    /// or has many messages needed to replay.
-    /// </summary>
-    public class ThrowOverflowExceptionStrategy : IStashOverflowStrategy
-    {
-        /// <summary>
-        /// TBD
-        /// </summary>
-        public static readonly ThrowOverflowExceptionStrategy Instance = new ThrowOverflowExceptionStrategy();
-
-        private ThrowOverflowExceptionStrategy() { }
-    }
-
-    /// <summary>
-    /// Reply to sender with predefined response, and discard the received message silently.
-    /// </summary>
-    public sealed class ReplyToStrategy : IStashOverflowStrategy
-    {
-        /// <summary>
-        /// TBD
-        /// </summary>
-        /// <param name="response">TBD</param>
-        public ReplyToStrategy(object response)
-        {
-            Response = response;
-        }
-
-        /// <summary>
-        /// The message replying to sender with
-        /// </summary>
-        public object Response { get; private set; }
-    }
-
-    /// <summary>
-    /// Implement this interface in order to configure the <see cref="IStashOverflowStrategy"/>
-    /// for the internal stash of the persistent actor.
-    /// An instance of this class must be instantiable using a no-args constructor.
-    /// </summary>
-    public interface IStashOverflowStrategyConfigurator
-    {
-        /// <summary>
-        /// TBD
-        /// </summary>
-        /// <param name="config">TBD</param>
-        /// <returns>TBD</returns>
-        IStashOverflowStrategy Create(Config config);
-    }
-
-    /// <summary>
-    /// TBD
-    /// </summary>
-    public sealed class ThrowExceptionConfigurator : IStashOverflowStrategyConfigurator
-    {
-        /// <summary>
-        /// TBD
-        /// </summary>
-        /// <param name="config">TBD</param>
-        /// <returns>TBD</returns>
-        public IStashOverflowStrategy Create(Config config)
-        {
-            return ThrowOverflowExceptionStrategy.Instance;
-        }
-    }
-
-    /// <summary>
-    /// TBD
-    /// </summary>
-    public sealed class DiscardConfigurator : IStashOverflowStrategyConfigurator
-    {
-        /// <summary>
-        /// TBD
-        /// </summary>
-        /// <param name="config">TBD</param>
-        /// <returns>TBD</returns>
-        public IStashOverflowStrategy Create(Config config)
-        {
-            return DiscardToDeadLetterStrategy.Instance;
-        }
-    }
-
-    /// <summary>
     /// Persistent actor - can be used to implement command or eventsourcing.
     /// </summary>
     public abstract class PersistentActor : Eventsourced
     {
-        /// <summary>
-        /// TBD
-        /// </summary>
-        /// <param name="message">TBD</param>
-        /// <returns>TBD</returns>
+        /// <inheritdoc />
         protected override bool Receive(object message)
         {
             return ReceiveCommand(message);
@@ -271,32 +30,20 @@ namespace Akka.Persistence
     /// </summary>
     public abstract class UntypedPersistentActor : Eventsourced
     {
-        /// <summary>
-        /// TBD
-        /// </summary>
-        /// <param name="message">TBD</param>
-        /// <returns>TBD</returns>
+        /// <inheritdoc />
         protected override bool Receive(object message)
         {
             return ReceiveCommand(message);
         }
 
-        /// <summary>
-        /// TBD
-        /// </summary>
-        /// <param name="message">TBD</param>
-        /// <returns>TBD</returns>
+        /// <inheritdoc />
         protected sealed override bool ReceiveCommand(object message)
         {
             OnCommand(message);
             return true;
         }
 
-        /// <summary>
-        /// TBD
-        /// </summary>
-        /// <param name="message">TBD</param>
-        /// <returns>TBD</returns>
+        /// <inheritdoc />
         protected sealed override bool ReceiveRecover(object message)
         {
             OnRecover(message);
@@ -308,6 +55,7 @@ namespace Akka.Persistence
         /// </summary>
         /// <param name="message">TBD</param>
         protected abstract void OnCommand(object message);
+
         /// <summary>
         /// TBD
         /// </summary>
@@ -338,7 +86,7 @@ namespace Akka.Persistence
         /// <summary>
         /// TBD
         /// </summary>
-        protected new static IUntypedActorContext Context { get { return (IUntypedActorContext)ActorBase.Context; } }
+        protected new static IUntypedActorContext Context => (IUntypedActorContext)ActorBase.Context;
     }
 
     /// <summary>
@@ -411,19 +159,13 @@ namespace Akka.Persistence
             base.BecomeStacked(m => ExecutePartialMessageHandler(m, newHandler));
         }
 
-        /// <summary>
-        /// TBD
-        /// </summary>
-        /// <param name="message">TBD</param>
+        /// <inheritdoc />
         protected sealed override void OnCommand(object message)
         {
             ExecutePartialMessageHandler(message, _partialReceiveCommand);
         }
 
-        /// <summary>
-        /// TBD
-        /// </summary>
-        /// <param name="message">TBD</param>
+        /// <inheritdoc />
         protected sealed override void OnRecover(object message)
         {
             ExecutePartialMessageHandler(message, _partialReceiveRecover);
@@ -435,8 +177,6 @@ namespace Akka.Persistence
             if (!wasHandled && _shouldUnhandle)
                 Unhandled(message);
         }
-
-        #region Recover helper methods
 
         private void EnsureMayConfigureRecoverHandlers()
         {
@@ -522,10 +262,6 @@ namespace Akka.Persistence
             _matchRecoverBuilders.Peek().MatchAny(handler);
         }
 
-        #endregion
-
-        #region Command helper methods
-        
         private void EnsureMayConfigureCommandHandlers()
         {
             if (_matchCommandBuilders.Count <= 0)
@@ -627,8 +363,5 @@ namespace Akka.Persistence
             var newHandler = BuildNewReceiveHandler(_matchCommandBuilders.Pop());
             return newHandler;
         }
-
-        #endregion
     }
 }
-
